djangotodo/requirements.txt
git+https://github.com/behave/behave.git
behave-django==0.5.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-extensions==1.9.1
djangorestframework==3.7.0
idna==2.6
parse==1.8.2
parse-type==0.4.2
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
